# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: cron-monitor
servers:
  - url: 'http://localhost:8081'
    description: '로컬 테스트 환경'
  - url: 'http://139.150.64.58:80'
    description: '배포 환경 로컬 접속'
  - url: 'http://10.7.27.11:80'
    description: '배포 환경 gcloud 접속'
  - url: 'http://182.162.142.151:8086'
    description: 'InfluxDB'
  - url: 'http://139.150.64.58:3306'
    description: 'MySQL DB 로컬 접속'
  - url: 'http://10.7.27.11:3306'
    description: 'MySQL DB gcloud 접속'
  - url: 'http://10.7.27.9:9092'
    description: '카프카 gcloud 접속'
  - url: 'http://182.162.142.151:9093'
    description: '카프카 로컬 접속'
  - url: 'http://localhost:6379'
    description: '테스트용 레디스 리프레시 토큰 DB'
  - url: '10.7.27.9:6379'
    description: '운영용 레디스 리프레시 토큰 DB'
paths:
...
  /notifications/users/{userId}/crons/{cronJobId}/webhooks:
    post:
      summary: '웹훅 추가'
      operationId: changeTeam
      tags:
        - webhook
      parameters:
        - $ref: '#/components/parameters/userIdPath'
        - $ref: '#/components/parameters/cronJobIdPath'
      requestBody:
        $ref: '#/components/requestBodies/PostNotificationsNotId'
      responses:
        '201':
          description: '정상적으로 변경됨'
        '400':
          description: 'Bad request'
    get:
      summary: '웹훅 리스트 조회'
      operationId: getWebhooks
      tags:
        - webhook
      parameters:
        - $ref: '#/components/parameters/userIdPath'
        - $ref: '#/components/parameters/cronJobIdPath'
      responses:
        '200':
          description: '정상적으로 조회됨'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'
        '400':
          description: 'Bad request'
    delete:
      summary: '알림에 대한 웹훅 일괄 삭제'
      operationId: deleteWebhooks
      tags:
        - webhook
      parameters:
        - $ref: '#/components/parameters/userIdPath'
        - $ref: '#/components/parameters/cronJobIdPath'
      responses:
        '204':
          description: '정상적으로 삭제됨'
        '400':
          description: 'Bad request'
  /notifications/users/{userId}/crons/{cronJobId}/webhooks/{webhookId}:
    patch:
      summary: '웹훅 수정'
      operationId: changeTeam
      tags:
        - webhook
      parameters:
        - $ref: '#/components/parameters/userIdPath'
        - $ref: '#/components/parameters/cronJobIdPath'
        - $ref: '#/components/parameters/webhookIdPath'
      requestBody:
        $ref: '#/components/requestBodies/PatchNotificationsNotId'
      responses:
        '201':
          description: '정상적으로 변경됨'
        '400':
          description: 'Bad request'
    delete:
      summary: '웹훅 개별 삭제'
      operationId: deleteWebhookById
      tags:
        - webhook
      parameters:
        - $ref: '#/components/parameters/userIdPath'
        - $ref: '#/components/parameters/cronJobIdPath'
        - $ref: '#/components/parameters/webhookIdPath'
      responses:
        '204':
          description: '정상적으로 삭제됨'
        '400':
          description: 'Bad request'
components:
  schemas:
...
    Webhook:
      description: '웹훅'
      type: object
      properties:
        url:
          type: string
...
  parameters:
    cronJobIdPath:
      name: cronJobId
      in: path
      description: '크론 job의 id'
      required: true
      schema:
        type: integer
...
    userIdPath:
      name: userId
      in: path
      description: '유저의 id'
      required: true
      schema:
        type: string
...
    webhookIdPath:
      name: webhookId
      in: path
      description: '웹훅의 id'
      required: true
      schema:
        type: string
  requestBodies:
...
    PostNotificationsNotId:
      description: '웹훅 추가'
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - url
            properties:
              url:
                type: string
    PatchNotificationsNotId:
      description: '웹훅 수정'
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - webhookId
              - url
            properties:
              pk:
                type: integer
              url:
                type: string